test(item-sliding): migrate icons
